Added login, account, user management and logs

+++ backend/app/api/job.py
@@ -1,26 +1,32 @@
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, Request
 from sqlalchemy.orm import Session
 from app.database import get_db
 from app.models.jobs import Job
 from app.schemas.mailer import MailerJobCreate, MailerJobOut, MailerJobUpdate
 from app.utils.responses import send_status_response
 
+from app.utils.security import Security
+
 router = APIRouter(prefix="/job", tags=["job"])
 
-@router.post("/", response_model=MailerJobOut)
-def create_mailer_job(data: MailerJobCreate, db: Session = Depends(get_db)):
-    job = Job(**data.dict())
+@router.post("", response_model=MailerJobOut)
+def create_mailer_job(request: Request, data: MailerJobCreate, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+    job = Job(**data.dict(), user_id=user.id)
     db.add(job)
     db.commit()
     db.refresh(job)
     return job
 
-@router.get("/", response_model=list[MailerJobOut])
-def list_mailer_jobs(db: Session = Depends(get_db)):
-    return db.query(Job).all()
+@router.get("", response_model=list[MailerJobOut])
+def list_mailer_jobs(request: Request, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+    return db.query(Job).filter(Job.user_id == user.id).all()
 
 @router.delete("/{job_id}")
-def delete_mailer_job(job_id: int, db: Session = Depends(get_db)):
+def delete_mailer_job(request: Request, job_id: str, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+
     job = db.query(Job).filter(Job.id == job_id).first()
     if not job:
         return send_status_response(
@@ -29,12 +35,22 @@ def delete_mailer_job(job_id: int, db: Session = Depends(get_db)):
             status=404,
             detail=f"Job with id {job_id} does not exist."
         )
+    if job.user_id != user.id:
+        return send_status_response(
+            code="UNAUTHORIZED",
+            message="Cannot delete: unauthorized",
+            status=403,
+            detail=f"User {user.id} is not the owner of job {job_id}."
+        )
+    
     db.delete(job)
     db.commit()
     return {"success": True}
 
 @router.get("/{job_id}", response_model=MailerJobOut)
-def get_mailer_job(job_id: int, db: Session = Depends(get_db)):
+def get_mailer_job(request: Request, job_id: str, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+
     job = db.query(Job).filter(Job.id == job_id).first()
     if not job:
         return send_status_response(
@@ -43,10 +59,20 @@ def get_mailer_job(job_id: int, db: Session = Depends(get_db)):
             status=404,
             detail=f"No job with id {job_id} exists."
         )
+    
+    if job.user_id != user.id:
+        return send_status_response(
+            code="UNAUTHORIZED",
+            message="Unauthorized access to job",
+            status=403,
+            detail=f"User {user.id} is not allowed to access job {job_id}."
+        )
     return job
 
 @router.put("/{job_id}", response_model=MailerJobOut)
-def update_mailer_job(job_id: int, data: MailerJobUpdate, db: Session = Depends(get_db)):
+def update_mailer_job(request: Request, job_id: str, data: MailerJobUpdate, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+
     job = db.query(Job).filter(Job.id == job_id).first()
     if not job:
         return send_status_response(
@@ -55,8 +81,44 @@ def update_mailer_job(job_id: int, data: MailerJobUpdate, db: Session = Depends(
             status=404,
             detail=f"Job with id {job_id} does not exist."
         )
+    
+    if job.user_id != user.id:
+        return send_status_response(
+            code="UNAUTHORIZED",
+            message="Unauthorized access to job",
+            status=403,
+            detail=f"User {user.id} is not allowed to update job {job_id}."
+        )
+    
     for key, value in data.dict().items():
         setattr(job, key, value)
     db.commit()
     db.refresh(job)
     return job
+
+@router.put("/{job_id}/status")
+def update_mailer_job(request: Request, job_id: str, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+
+    job = db.query(Job).filter(Job.id == job_id).first()
+    if not job:
+        return send_status_response(
+            code="UPDATE_FAILED",
+            message="Cannot update: job not found",
+            status=404,
+            detail=f"Job with id {job_id} does not exist."
+        )
+    
+    if job.user_id != user.id:
+        return send_status_response(
+            code="UNAUTHORIZED",
+            message="Unauthorized access to job",
+            status=403,
+            detail=f"User {user.id} is not allowed to update job {job_id}."
+        )
+
+    job.is_active = not job.is_active
+    db.commit()
+    db.refresh(job)
+
+    return {"is_active": job.is_active}
